Include example of Linux thread implementation
This example is in listing 4.15 of section 4.5.  It is supposed to show
that threads are really processes in GNU/Linux and that each has its own
processes ID.  But running on a kernel version 4.17.2 shows that both
threads has the same PID.  Executing the program in the background and
then executing `ps x` shows that there is only one process running.  If
instead of executing `ps x` you execute `top` (or `htop`) you can see
that there are two processes with different PIDs, eventhough each thread
prints the same PID when executing `getpid()`.

Maybe this happened because the book uses an old kernel version.